feat: Add hunting query for detecting processes reading a large number of unique files
This commit adds a new hunting query to detect processes that read a large number of unique files in a short period of time. The query helps identify potential attackers using tools to enumerate files on a system and scan for sensitive information. The query is based on the technique T1005 from MITRE ATT&CK and is authored by Linus.